Fix fsync bug that prevented editing teacher files (#24)
When using `fysync` to sync data to disk the function `fsync` was called on the
file object, which calls `File.flush`. This function returns a serialized code
object when it got such an object from the server. However, the `File.fsync`
function didn't return `File.flush`, so the updated file `id` was never
saved. This PR fixes this issue.